Moved generation classes to "generation" package
