+++ client/src/test/scala/it/cwmp/client/controller/game/CellWorldGenerationStrategyTest.scala
@@ -1,5 +1,6 @@
 package it.cwmp.client.controller.game
 
+import it.cwmp.client.controller.game.generation.CellWorldGenerationStrategy
 import it.cwmp.client.model.game.impl.Cell
 import it.cwmp.model.User
 import org.scalatest.FunSpec
